+++ tests/backend_test.py
@@ -23,6 +23,7 @@ from pytket.extensions.braket import BraketBackend
 from pytket.architecture import FullyConnected
 from pytket.circuit import Circuit, OpType, Qubit, Bit
 from pytket.pauli import Pauli, QubitPauliString
+from pytket.passes import SequencePass, BasePass
 from pytket.utils.expectations import (
     get_pauli_expectation_value,
     get_operator_expectation_value,
@@ -633,3 +634,50 @@ def test_multiple_indices_rigetti(authenticated_braket_backend: BraketBackend) -
     c1 = b.get_compiled_circuit(c)
     h = b.process_circuit(c1, 100)
     b.cancel(h)
+
+
+# Helper function used for testing serialization
+# Both local and remote backends are tested.
+def run_serialization_test(backend: BraketBackend) -> None:
+    for opt_level in range(3):
+        default_pass = backend.default_compilation_pass(opt_level)
+        original_pass_dict = default_pass.to_dict()
+        reconstructed_pass = BasePass.from_dict(original_pass_dict)
+        assert isinstance(reconstructed_pass, SequencePass)
+        assert original_pass_dict == reconstructed_pass.to_dict()
+
+
+def test_local_backend_pass_serialization() -> None:
+    local_backend = BraketBackend(local=True)
+    run_serialization_test(local_backend)
+
+
+@pytest.mark.skipif(skip_remote_tests, reason=REASON)
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [
+        {
+            "device_type": "qpu",
+            "region": "us-east-1",
+            "provider": "ionq",
+            "device": "Aria-1",
+        },
+        {
+            "device_type": "qpu",
+            "provider": "rigetti",
+            "device": "Ankaa-3",
+            "region": "us-west-1",
+        },
+        {
+            "device_type": "qpu",
+            "provider": "iqm",
+            "device": "Garnet",
+            "region": "eu-north-1",
+        },
+    ],
+    indirect=True,
+)
+def test_remote_backend_pass_serialization(
+    authenticated_braket_backend: BraketBackend,
+) -> None:
+    run_serialization_test(authenticated_braket_backend)